files_elasticsearch/backend/FilesContentSearch/requirements.txt
asgiref==3.3.4
certifi==2022.6.15
charset-normalizer==2.1.0
Django==3.2.4
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
idna==3.3
pytz==2021.1
requests==2.28.1
sqlparse==0.4.1
tika==1.24
typing-extensions==3.10.0.0
urllib3==1.26.11
